requirements.txt
argparse==1.4.0
numpy==1.21.0
pathlib==1.0.1
h5py==3.6.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cu111
torchvision==0.8.2+cu110
torchaudio==0.8.0
pyYAML==6.0
matplotlib==3.5.1
scipy==1.8.0
einops==0.4.1
sklearn==0.0
seaborn==0.11.2